+++ packages/plugin/src/Library/DataObjects/FormRenderObject/AbstractFormRenderObject.php
@@ -3,24 +3,31 @@
 namespace Solspace\Freeform\Library\DataObjects\FormRenderObject;
 
 use Solspace\Commons\Helpers\StringHelper;
+use Solspace\Freeform\Events\Forms\RegisterRenderObjectOptionsEvent;
+use yii\base\Event;
 
 abstract class AbstractFormRenderObject implements FormRenderObjectInterface
 {
+    const EVENT_REGISTER_OPTIONS = 'registerOptions';
+
+    /** @var array */
+    protected $options;
+
     /** @var mixed */
     private $value;
 
     /** @var array */
     private $replacements;
 
-    /**
-     * AbstractFormRenderObject constructor.
-     *
-     * @param mixed $value
-     */
-    public function __construct($value, array $replacements = [])
+    public function __construct($value, array $replacements = [], $options = [])
     {
         $this->value = $value;
         $this->replacements = $replacements;
+
+        $event = new RegisterRenderObjectOptionsEvent($options);
+        Event::trigger(static::class, self::EVENT_REGISTER_OPTIONS, $event);
+
+        $this->options = $event->getOptions();
     }
 
     /**